Switch to building and linking CSS and JS files individually
With HTTP/2 requests to the same domain are multiplexed on the same connection.
This means that there is no longer a penalty for having lots of small assets as
they can be requested in parallel and cached individually.

I'm planning on putting Nginx with HTTP/2 enabled in front of the app, so the
best thing to do is to compile and serve each asset individually.